requirements.txt
Django>=4.2,<5.0
requests
python-dotenv
google-genai